+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email "sawanoboriyu@higanworks.com"
 license          "MIT"
 description      "Installs/Configures mongodb-10gen"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.3.0"
+version          "0.4.0"
 depends          "apt"
 supports         "ubuntu"
 supports         "debian"